cleaned up firebase context
