+++ src/Atoms/Firebase.jsx
@@ -1,3 +1,4 @@
+import React from 'react';
 import 'firebase/auth';
 
 import app from 'firebase/app';
@@ -29,4 +30,13 @@ class Firebase {
   doPasswordUpdate = password => this.auth.currentUser.updatePassword(password);
 }
 
+const FirebaseContext = React.createContext(null);
+
+const withFirebase = Component => props => (
+  <FirebaseContext.Consumer>
+    {firebase => <Component {...props} firebase={firebase} />}
+  </FirebaseContext.Consumer>
+);
+
+export { FirebaseContext, withFirebase };
 export default Firebase;
