--- src/Atoms/Firebase/index.js
@@ -1,7 +1,0 @@
-import FirebaseContext, { withFirebase } from './context';
-
-import Firebase from './firebase';
-
-export default Firebase;
-
-export { FirebaseContext, withFirebase };
